introspection: rename package, rm ListenAddr() from interface. (#111)
