+++ host/host.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/connmgr"
 	"github.com/libp2p/go-libp2p-core/event"
-	"github.com/libp2p/go-libp2p-core/introspect"
+	"github.com/libp2p/go-libp2p-core/introspection"
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/libp2p/go-libp2p-core/peerstore"
@@ -79,9 +79,9 @@ type Host interface {
 // introspectable, that is, that expose an introspection server.
 type IntrospectableHost interface {
 
-	// Introspector returns the Introspector instance, with which the caller
-	// can:
+	// Introspector returns the introspection.Introspector instance, with which
+	// the caller can:
 	//  - register data providers.
 	//  - fetch introspection data.
-	Introspector() introspect.Introspector
+	Introspector() introspection.Introspector
 }
