+++ introspection/doc.go
@@ -1,4 +1,4 @@
-// Package introspect is EXPERIMENTAL. It is subject to heavy change, and it
+// Package introspection is EXPERIMENTAL. It is subject to heavy change, and it
 // WILL change. For now, it is the simplest implementation to power the
 // proof-of-concept of the libp2p introspection framework.
 //
@@ -6,4 +6,4 @@
 // of go-libp2p. It holds the introspection data schema, and the primitives that
 // allow subsystems to register data providers, and clients to fetch the current
 // state of the system.
-package introspect
+package introspection
